Pretty Massive Sprite Refactor
Refactor how sprites/textures are loaded into the game. The game now
reads an atlas.json which is used to populate a hash table that maps a
sprite key to a set of texture offsets. This has done the important job
of decoupling the glyps from explicit character mappings. Now I can
work on UI element sprites :-)