Rudimentary Save/Load/Gen commands
Add commands to save/load/gen worlds. Quite rudimentary, slow and basic.
Probably lots of unnecessary clearing/wiping of state. Also need to
figure out a way to make it not stop-the-world for a few seconds :(

Figured out its slow because of the reading/writing of data - not so
much the lighting calculations...