Text/Input
Changed the way we render text. Glyphs are now wrapped in a
re-usable/pooled ConfiguredGlyph object that contains information about
color and fontdecoration. Created a TextFragment object which holds a
list of these ConfiguredGlyphs. We can push fragments of text to this
TextFragment object (even single characters). This allows us to have
strings of text with different colors/styles.

Leveraged this to soup up the DebugInformation and also upgraded the
TextInputComponent to show a black highlighting cursor using this
facility. Also added more keyboard controls to the TextInputControl (L,
R, HOME, END, DEL)